Move keychainDel to darwin-only file to fix unused lint error
